backend/requirements.txt
# ── gRPC core ──
grpcio==1.68.1
grpcio-tools==1.68.1
grpcio-reflection==1.68.1
protobuf==5.29.2

# ── HTTP Gateway (SSE streaming for Render/Vercel compatibility) ──
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.3

# ── Testing ──
pytest==8.3.4
pytest-asyncio==0.24.0
grpcio-testing==1.68.1
httpx==0.28.1

# ── Utilities ──
python-dotenv==1.0.1
